requirements.txt
azure-core==1.32.0
azure-storage-blob==12.8.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
gunicorn==20.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msrest==0.7.1
mysql-connector-python==9.1.0
oauthlib==3.2.2
pycparser==2.22
pyodbc==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
waitress==2.1.1

